Move components to subpackage
